% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/template.R
\name{templateOutput}
\alias{templateOutput}
\title{Widget output function for use in Shiny}
\usage{
templateOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

